+++ pytsite/__init__.py
@@ -144,9 +144,9 @@ def _init():
     from . import _eh
     events.listen('pytsite.update', _eh.update)
 
-    # Initialize theme
+    # Initialize default theme
     from pytsite import theme
-    theme.load()
+    theme.get().load()
 
 
 _init()