+++ pytsite/lang/_eh.py
@@ -9,9 +9,8 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
-def assetman_build(console_notify: bool):
-    if console_notify:
-        _console.print_info(_lang.t('pytsite.assetman@compiling_translations'))
+def assetman_build():
+    _console.print_info(_lang.t('pytsite.assetman@compiling_translations'))
 
     translations = {}
     for lang_code in _lang.langs():