+++ pytsite/tpl/__init__.py
@@ -22,7 +22,7 @@ def _get_tpl_path(tpl: str) -> str:
         tpl = '$theme@' + tpl
 
     if '$theme' in tpl:
-        tpl = tpl.replace('$theme', _theme.get_current())
+        tpl = tpl.replace('$theme', _theme.get().package_name)
 
     package_name, tpl_name = tpl.split('@')[:2]